interaction motion factor fix
using setmotion factor command to reduce the sensitivity of the interaction. this will prevent the fast motion of the camera zoom in and out .
Zoom in and zoom out set motion factor as 1 for normal speed zoom in and zoom out with the camera.zoom(0.99) for example.
Mouse move , keyboard and room interaction will stick to 8-10 for the motionfactor for default motion.
So that include my fixes for the interaction. Also i include collision fliter to detect collision between two objects.