run_marking.sh in progress.
According to my report , I found some discovery to change from my site.
output for point , if point ------ done . It is indicate as counter to send to the ui .
Once the ui reach the full marking counter/rows . It will indicate as wall done , otherwise I will include an error.
Right now I cannot test because the machine has some malfunction issues based on position and rotation.
Once the robot was fixed , I will continue on testing the error and the point counter .